# docs/openapi/fortuna.json
{
  "components": {
    "schemas": {
...
      "InstallmentCommandOutput": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "format": "double",
            "type": "number"
          },
          "appliedRate": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "currencyCode": {
            "nullable": true,
            "type": "string"
          },
          "isLateArriving": {
            "type": "boolean"
          },
          "number": {
            "format": "int32",
            "type": "integer"
          },
          "occurredOn": {
            "format": "date",
            "type": "string"
          },
          "originalAmount": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "originalCurrencyCode": {
            "nullable": true,
            "type": "string"
          },
          "rateDate": {
            "format": "date",
            "nullable": true,
            "type": "string"
          },
          "statementId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "transactionId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "type": "object"
      },
      "InstallmentOutput": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "format": "double",
            "type": "number"
          },
          "appliedRate": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "currencyCode": {
            "nullable": true,
            "type": "string"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "isLateArriving": {
            "type": "boolean"
          },
          "number": {
            "format": "int32",
            "type": "integer"
          },
          "occurredOn": {
            "format": "date",
            "type": "string"
          },
          "originalAmount": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "originalCurrencyCode": {
            "nullable": true,
            "type": "string"
          },
          "rateDate": {
            "format": "date",
            "nullable": true,
            "type": "string"
          },
          "statementId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "transactionId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "type": "object"
      },
      "InstallmentPlanLifecycleCommandOutput": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "type": "object"
      },
      "InstallmentPlanLifecycleCommandOutputDataOutput": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/InstallmentPlanLifecycleCommandOutput"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "messages": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "success": {
            "readOnly": true,
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "InstallmentPlanOutput": {
        "additionalProperties": false,
        "properties": {
          "appliedRate": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "creditCardId": {
            "format": "uuid",
            "type": "string"
          },
          "currencyCode": {
            "nullable": true,
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "installmentCount": {
            "format": "int32",
            "type": "integer"
          },
          "installments": {
            "items": {
              "$ref": "#/components/schemas/InstallmentOutput"
            },
            "nullable": true,
            "type": "array"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "originalCurrencyCode": {
            "nullable": true,
            "type": "string"
          },
          "originalTotalAmount": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "purchasedOn": {
            "format": "date",
            "type": "string"
          },
          "rateDate": {
            "format": "date",
            "nullable": true,
            "type": "string"
          },
          "totalAmount": {
            "format": "double",
            "type": "number"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "InstallmentPlanOutputDataOutput": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/InstallmentPlanOutput"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "messages": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "success": {
            "readOnly": true,
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "RecordInstallmentPlanCommand": {
        "additionalProperties": false,
        "properties": {
          "categoryId": {
            "format": "uuid",
            "type": "string"
          },
          "counterparty": {
            "nullable": true,
            "type": "string"
          },
          "creditCardId": {
            "format": "uuid",
            "type": "string"
          },
          "currencyCode": {
            "nullable": true,
            "type": "string"
          },
          "installmentCount": {
            "format": "int32",
            "type": "integer"
          },
          "ownerId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "purchasedOn": {
            "format": "date",
            "type": "string"
          },
          "totalAmount": {
            "format": "double",
            "type": "number"
          }
        },
        "type": "object"
      },
      "RecordInstallmentPlanCommandOutput": {
        "additionalProperties": false,
        "properties": {
          "appliedRate": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "creditCardId": {
            "format": "uuid",
            "type": "string"
          },
          "currencyCode": {
            "nullable": true,
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "installmentCount": {
            "format": "int32",
            "type": "integer"
          },
          "installments": {
            "items": {
              "$ref": "#/components/schemas/InstallmentCommandOutput"
            },
            "nullable": true,
            "type": "array"
          },
          "originalCurrencyCode": {
            "nullable": true,
            "type": "string"
          },
          "originalTotalAmount": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "purchasedOn": {
            "format": "date",
            "type": "string"
          },
          "rateDate": {
            "format": "date",
            "nullable": true,
            "type": "string"
          },
          "totalAmount": {
            "format": "double",
            "type": "number"
          }
        },
        "type": "object"
      },
      "RecordInstallmentPlanCommandOutputDataOutput": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RecordInstallmentPlanCommandOutput"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "messages": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "success": {
            "readOnly": true,
            "type": "boolean"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "bearerFormat": "JWT",
        "description": "Enter the Heimdall JWT bearer token.",
        "scheme": "Bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Fortuna API",
    "version": "v1"
  },
  "openapi": "3.0.4",
  "paths": {
...
    "/api/installment-plans": {
      "post": {
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RecordInstallmentPlanCommand"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordInstallmentPlanCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordInstallmentPlanCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordInstallmentPlanCommandOutputDataOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordInstallmentPlanCommandOutputDataOutput"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RecordInstallmentPlanCommandOutputDataOutput"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "InstallmentPlans"
        ]
      }
    },
    "/api/installment-plans/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallmentPlanLifecycleCommandOutputDataOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallmentPlanLifecycleCommandOutputDataOutput"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InstallmentPlanLifecycleCommandOutputDataOutput"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "InstallmentPlans"
        ]
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "includeDeleted",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallmentPlanOutputDataOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallmentPlanOutputDataOutput"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InstallmentPlanOutputDataOutput"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "InstallmentPlans"
        ]
      }
    },
    "/api/installment-plans/{id}/restore": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallmentPlanLifecycleCommandOutputDataOutput"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallmentPlanLifecycleCommandOutputDataOutput"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InstallmentPlanLifecycleCommandOutputDataOutput"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "InstallmentPlans"
        ]
      }
    },
...
  },
  "security": [
    {}
  ],
...
}